[BugFix] Fix year calculation edge case for scope computation (Calendar functions) (#1911)
## Why I'm doing:

@philiprebeschini reported a diff between Celostar and Saola of
WORKDAYS_BETWEEN in https://github.com/celonis/celostar/pull/15725. In
the case, an out of scope timestamp does not return NULL.

## What I'm doing:

This PR fixed the bug in `get_year` function used by `set_scope` method
where the last millisecond of a year is covered would incorrectly return
the next year (it overestimates the scope of a calendar). Calendar scope
is used in some calendar functions, when the input timestamp is out of
the scope, it returns NULL.

The repro SQL is
https://gist.github.com/zhangyaqi1989/0bc1d75b1dda723a5963a0283f17a752
(confirmed it is fixed by this pr).

This PR also adds a unit test.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5