Add dev images for distributed compilation and devcontainer for ARM - CELOSTAR-991 (#1957)
This adds two Dockerfiles so that distributed compilation on Mac (ARM)
works:
- `dev-env-arm-homcc`: This is used locally for a Linux devcontainer. It
essentially takes a `starrocks-dev-env` as base image and installs
[homcc](https://github.com/celonis/homcc) (our distributed compiler) and
other useful build tools.
- `dev-env-x86-to-arm-cross-compilation`: Similarly to the client image
it also reuses a base `starrocks-dev-env` image but installs everything
for cross-compiling on x86 to ARM. This image runs on the homcc servers.

Then this PR also adds a [devcontainer](https://containers.dev/) for
exactly this use case. On Mac with ARM you can use this dev container
and distributed compilation will work out-of-the-box. For this I added
some CMake presets so one does not have to use SR's `build.sh` script
any more. For now I added a preset for building SR with debug + tests
and SR with release without tests.
I also included basic config for VS code, but note the `devcontainer`
also works for CLion or if you bare-metal wan tot execute `build.sh`.