ci: Check clang-format for `celonis` subdirs (#1993)
We should enforce `clang-format` at least on our files (SR has too many
uncompliant files unfortunately). This PR adds a CI check that enforces
formatting of

- `be/src/exprs/celonis`
- `be/test/exprs/celonis`
- `be/src/bench/celonis`

I formatted all uncompliant files so that we are compliant now. (hence
the large diff)

The main change is in
[presubmit_test.sh](https://github.com/celonis/celostar-starrocks/pull/1993/files#diff-1ea5a2cc5aba69427057a1dbd3892064126be1de6789ba3c9596fb87d09299b9)

Note: We need `clang-format` in the dev env, hence
https://github.com/celonis/celostar-starrocks/pull/1996 ideally should
be merged and then this PR should be rebased.