[Enhancement] Improve the performance of CELONIS_SANITIZE_INVALID_UTF8 (~ 2.38x speedup for valid string) (#2049)
## Why I'm doing:

In Celostar, CELONIS_SANITIZE_INVALID_UTF8 is applied to all the string
columns ingested. So its performance matters.

## What I'm doing:

This PR adds a fast path to it (when the input string is valid) and uses
simdutf::validate_utf8 in the fast path. On average, it achieves 2.38x
speed up. The longer the input string, the more speed up it achieves
(because simdutf::validate_utf8 uses SIMD).

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5