[BugFix] Fix NODES check in match_activities incorrectly counting duplicate activities (#2052)
## Why I'm doing:

When checking for the presence of a set of required activities (NODES),
the logic attempted to predict if the current activity would complete
the set by checking nodes_seen.size() + 1 == nodes.size(). However, this
check failed to account for cases where the current activity was already
in nodes_seen. This caused cases with duplicate activities (e.g., ['A',
'A'] matching against {'A', 'B'}) to return a false positive.

This bug is the root cause of
https://celonis.atlassian.net/browse/CBE-49551.

## What I'm doing:

The logic has been updated to insert the current activity into the
nodes_seen set first (handling uniqueness automatically) and then check
if the size requirement is met.

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5